Move postcode handling to postal address
This prevents circular dependencies which meant that previously we’d
need to do imports at the method, not module level.
